+++ src/NodeAnalyzer/ParamAnalyzer.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Rector\Core\NodeAnalyzer;
 
 use PhpParser\Node;
+use PhpParser\Node\Expr\Assign;
 use PhpParser\Node\Expr\CallLike;
 use PhpParser\Node\Expr\Closure;
 use PhpParser\Node\Expr\ConstFetch;
@@ -19,8 +20,10 @@ use PhpParser\Node\Stmt\Function_;
 use PhpParser\NodeTraverser;
 use Rector\Core\NodeManipulator\FuncCallManipulator;
 use Rector\Core\PhpParser\Comparing\NodeComparator;
+use Rector\Core\PhpParser\Node\BetterNodeFinder;
 use Rector\Core\PhpParser\Node\Value\ValueResolver;
 use Rector\NodeNameResolver\NodeNameResolver;
+use Rector\NodeTypeResolver\Node\AttributeKey;
 use Symplify\Astral\NodeTraverser\SimpleCallableNodeTraverser;
 
 final class ParamAnalyzer
@@ -31,6 +34,7 @@ final class ParamAnalyzer
         private readonly NodeNameResolver $nodeNameResolver,
         private readonly FuncCallManipulator $funcCallManipulator,
         private readonly SimpleCallableNodeTraverser $simpleCallableNodeTraverser,
+        private readonly BetterNodeFinder $betterNodeFinder
     ) {
     }
 
@@ -137,5 +141,29 @@ final class ParamAnalyzer
 
         $arguments = $this->funcCallManipulator->extractArgumentsFromCompactFuncCalls([$node]);
         return $this->nodeNameResolver->isNames($param, $arguments);
+    }
+
+    public function isParamReassign(Param $param): bool
+    {
+        $classMethod = $param->getAttribute(AttributeKey::PARENT_NODE);
+
+        if (! $classMethod instanceof ClassMethod) {
+            return false;
+        }
+
+        $paramName = (string) $this->nodeNameResolver->getName($param->var);
+        return (bool) $this->betterNodeFinder->findFirstInFunctionLikeScoped($classMethod, function (Node $node) use (
+            $paramName
+        ): bool {
+            if (! $node instanceof Assign) {
+                return false;
+            }
+
+            if (! $node->var instanceof Variable) {
+                return false;
+            }
+
+            return $this->nodeNameResolver->isName($node->var, $paramName);
+        });
     }
 }
